# Conflicts:
#	src/wp43s/statusBar.c
#	src/wp43s/wp43s.h

+++ src/wp43s/integers.c
@@ -113,7 +113,7 @@ void longIntegerMultiply(longInteger_t opY, longInteger_t opX, longInteger_t res
   else {
     displayCalcErrorMessage(longIntegerSign(opY) == longIntegerSign(opX) ? ERROR_OVERFLOW_PLUS_INF : ERROR_OVERFLOW_MINUS_INF, ERR_REGISTER_LINE, REGISTER_X);
     #if(EXTRA_INFO_ON_CALC_ERROR == 1)
-      sprintf(errorMessage, "Multiplying this 2 values (%" PRIu64 " bits " STD_CROSS " %" PRIu64 " bits) would result in a value exceeding %" PRId16 " bits!", (uint64_t)longIntegerBits(opY), (uint64_t)longIntegerBits(opX), MAX_LONG_INTEGER_SIZE_IN_BITS);
+      sprintf(errorMessage, "Multiplying this 2 values (%" PRIu64 " bits " STD_CROSS " %" PRIu64 " bits) would result in a value exceeding %" PRId16 " bits!", (uint64_t)longIntegerBits(opY), (uint64_t)longIntegerBits(opX), (uint16_t)MAX_LONG_INTEGER_SIZE_IN_BITS);
       longIntegerToAllocatedString(opY, tmpString, TMP_STR_LENGTH / 2);
       longIntegerToAllocatedString(opX, tmpString + TMP_STR_LENGTH / 2, TMP_STR_LENGTH / 2);
       moreInfoOnError("In function longIntegerMultiply:", errorMessage, tmpString, tmpString + TMP_STR_LENGTH / 2);
@@ -130,7 +130,7 @@ void longIntegerSquare(longInteger_t op, longInteger_t result) {
   else {
     displayCalcErrorMessage(ERROR_OVERFLOW_PLUS_INF, ERR_REGISTER_LINE, REGISTER_X);
     #if(EXTRA_INFO_ON_CALC_ERROR == 1)
-      sprintf(errorMessage, "Squaring this value (%" PRIu64 " bits) would result in a value exceeding %" PRId16 " bits!", (uint64_t)longIntegerBits(op), MAX_LONG_INTEGER_SIZE_IN_BITS);
+      sprintf(errorMessage, "Squaring this value (%" PRIu64 " bits) would result in a value exceeding %" PRId16 " bits!", (uint64_t)longIntegerBits(op), (uint16_t)MAX_LONG_INTEGER_SIZE_IN_BITS);
       longIntegerToAllocatedString(op, tmpString, TMP_STR_LENGTH);
       moreInfoOnError("In function longIntegerSquare:", errorMessage, tmpString, NULL);
     #endif // (EXTRA_INFO_ON_CALC_ERROR == 1)
@@ -146,7 +146,7 @@ void longIntegerAdd(longInteger_t opY, longInteger_t opX, longInteger_t result) 
   else {
     displayCalcErrorMessage(longIntegerSign(opY) == 0 ? ERROR_OVERFLOW_PLUS_INF : ERROR_OVERFLOW_MINUS_INF, ERR_REGISTER_LINE, REGISTER_X);
     #if(EXTRA_INFO_ON_CALC_ERROR == 1)
-      sprintf(errorMessage, "Adding this 2 values (%" PRIu64 " bits " STD_CROSS " %" PRIu64 " bits) would result in a value exceeding %" PRId16 " bits!", (uint64_t)longIntegerBits(opY), (uint64_t)longIntegerBits(opX), MAX_LONG_INTEGER_SIZE_IN_BITS);
+      sprintf(errorMessage, "Adding this 2 values (%" PRIu64 " bits " STD_CROSS " %" PRIu64 " bits) would result in a value exceeding %" PRId16 " bits!", (uint64_t)longIntegerBits(opY), (uint64_t)longIntegerBits(opX), (uint16_t)MAX_LONG_INTEGER_SIZE_IN_BITS);
       longIntegerToAllocatedString(opY, tmpString, TMP_STR_LENGTH / 2);
       longIntegerToAllocatedString(opX, tmpString + TMP_STR_LENGTH / 2, TMP_STR_LENGTH / 2);
       moreInfoOnError("In function longIntegerAdd:", errorMessage, tmpString, tmpString + TMP_STR_LENGTH / 2);
@@ -163,7 +163,7 @@ void longIntegerSubtract(longInteger_t opY, longInteger_t opX, longInteger_t res
   else {
     displayCalcErrorMessage(longIntegerSign(opY) == 0 ? ERROR_OVERFLOW_PLUS_INF : ERROR_OVERFLOW_MINUS_INF, ERR_REGISTER_LINE, REGISTER_X);
     #if(EXTRA_INFO_ON_CALC_ERROR == 1)
-      sprintf(errorMessage, "Subtracting this 2 values (%" PRIu64 " bits " STD_CROSS " %" PRIu64 " bits) would result in a value exceeding %" PRId16 " bits!", (uint64_t)longIntegerBits(opY), (uint64_t)longIntegerBits(opX), MAX_LONG_INTEGER_SIZE_IN_BITS);
+      sprintf(errorMessage, "Subtracting this 2 values (%" PRIu64 " bits " STD_CROSS " %" PRIu64 " bits) would result in a value exceeding %" PRId16 " bits!", (uint64_t)longIntegerBits(opY), (uint64_t)longIntegerBits(opX), (uint16_t)MAX_LONG_INTEGER_SIZE_IN_BITS);
       longIntegerToAllocatedString(opY, tmpString, TMP_STR_LENGTH / 2);
       longIntegerToAllocatedString(opX, tmpString + TMP_STR_LENGTH / 2, TMP_STR_LENGTH / 2);
       moreInfoOnError("In function longIntegerSubtract:", errorMessage, tmpString, tmpString + TMP_STR_LENGTH / 2);
